add more tests + some refactoring
