+++ compatibility/compatibility-vec12to20/src/test/java/com/foursoft/harness/vec12to20/wrapper/AbstractBaseWrapperTest.java
@@ -23,26 +23,15 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.wrapper;
+package com.foursoft.harness.vec12to20.wrapper;
 
-import com.foursoft.harness.TestFiles;
-import com.foursoft.harness.compatibility.vec12to20.util.DefaultVecReader;
-import com.foursoft.harness.vec.v2x.VecContent;
-import org.junit.jupiter.api.Test;
+import com.foursoft.harness.compatibility.core.CompatibilityContext;
+import com.foursoft.harness.compatibility.vec12to20.Vec12XTo20XCompatibilityWrapper;
 
-import java.io.IOException;
-import java.io.InputStream;
+public abstract class AbstractBaseWrapperTest {
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-class Vec12To20ContentWrapperTest extends AbstractBaseWrapperTest {
-
-    @Test
-    void invokeTest() throws IOException {
-        final InputStream inputOriginal = TestFiles.getInputStream(TestFiles.OLD_BEETLE_V12X);
-        final VecContent originalContent = DefaultVecReader.read(inputOriginal, "test");
-        assertThat(originalContent).isNotNull();
-        assertThat(originalContent.getVecVersion()).isEqualTo("2.0.2");
+    protected static CompatibilityContext get12To20Context() {
+        return new Vec12XTo20XCompatibilityWrapper().getContext();
     }
 
-} 
+}
