+++ compatibility/compatibility-vec12to20/src/test/java/com/foursoft/harness/vec12to20/wrapper/Vec12To20DocumentVersionWrapperTest.java
@@ -23,40 +23,37 @@
  * THE SOFTWARE.
  * =========================LICENSE_END==================================
  */
-package com.foursoft.harness.wrapper;
+package com.foursoft.harness.vec12to20.wrapper;
 
 import com.foursoft.harness.TestFiles;
 import com.foursoft.harness.compatibility.vec12to20.util.DefaultVecReader;
-import com.foursoft.harness.vec.common.util.StreamUtils;
 import com.foursoft.harness.vec.v2x.VecContent;
-import com.foursoft.harness.vec.v2x.VecExtendableElement;
-import com.foursoft.harness.vec.v2x.VecLocalizedString;
-import com.foursoft.harness.vec.v2x.VecPartVersion;
+import com.foursoft.harness.vec.v2x.VecDocumentVersion;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Optional;
+import java.math.BigInteger;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class Vec12To20PartVersionWrapperTest extends AbstractBaseWrapperTest {
+class Vec12To20DocumentVersionWrapperTest extends AbstractBaseWrapperTest {
 
     @Test
     void invokeTest() throws IOException {
-        final InputStream inputOriginal = TestFiles.getInputStream(TestFiles.OLD_BEETLE_V12X);
-        final VecContent originalContent = DefaultVecReader.read(inputOriginal, "test");
-        final Optional<VecPartVersion> vecPartVersion = originalContent.getPartVersions().stream()
-                .filter(pv -> !pv.getPreferredUseCases().isEmpty())
-                .sorted(Comparator.comparing(VecExtendableElement::getXmlId))
-                .collect(StreamUtils.findOneOrNone());
+        try (final InputStream inputOriginal = TestFiles.getInputStream(TestFiles.OLD_BEETLE_V12X)) {
+            final VecContent originalContent = DefaultVecReader.read(inputOriginal, "test");
+            assertThat(originalContent).isNotNull();
 
-        assertThat(vecPartVersion).isPresent();
-        final List<VecLocalizedString> preferredUseCases = vecPartVersion.get().getPreferredUseCases();
-        assertThat(preferredUseCases).hasSize(1);
-        assertThat(preferredUseCases.get(0).getValue()).isEqualTo("Normal Connector");
+            final VecDocumentVersion vecDocumentVersion = originalContent.getDocumentVersions().stream()
+                    .filter(d -> !d.getSheetOrChapters().isEmpty())
+                    .toList().get(0);
+
+            final BigInteger numberOfSheets = vecDocumentVersion.getNumberOfSheets();
+            assertThat(numberOfSheets).isEqualTo(1);
+
+            vecDocumentVersion.setNumberOfSheets(new BigInteger("5"));
+            assertThat(vecDocumentVersion.getNumberOfSheets()).isEqualTo(5);
+        }
     }
-
 }
